validate password

+++ src/controllers/pharmacyController.js
@@ -89,11 +89,18 @@ export const editPharmacy = async (req, res) => {
 
 export const deletePharmacy = async (req, res) => {
   const { userId, email, role } = req.user;
-
+  const password = req.body.password;
   try {
+    if (!password) {
+      return res.status(400).json({ message: "Password is required" });
+    }
     const user = await validateUser(userId, email, role);
     if (!user) {
       return res.status(404).json({ message: "User not found" });
+    }
+    const isMatch = await bcrypt.compare(password, user.password);
+    if (!isMatch) {
+      return res.status(401).json({ message: "Invalid password" });
     }
 
     await Pharmacy.findByIdAndDelete(userId);
